feat: add request on projectForm submittion and get all projects
