+++ src/pages/AdminPanel/AdminProjects/AddProject/AddProject.jsx
@@ -1,6 +1,3 @@
-import React from "react";
-import { useFormik } from "formik";
-import * as Yup from "yup";
 import { CustomSelect } from "../../../../components/AdminPanel/Filters/CustomSelect/CustomSelect";
 import { DateSelect } from "../../../../components/AdminPanel/Filters/DateSelect/DateSelect";
 import ImageInput from "../../../../components/AdminPanel/ImageInput/ImageInput";
@@ -9,50 +6,13 @@ import AdminButton from "../../../../components/AdminPanel/UI/Button/AdminButton
 import { Tooltip } from "../Tooltip/Tooltip";
 import styles from "./AddProject.module.css";
 import AdminHeader from "../../../../components/AdminPanel/Header/AdminHeader";
-
-const MockedOptions = [
-  "Тунельний політ",
-  "Банджі джампінг",
-  "Стрибки з парашутом",
-  "Стежка у хмарах",
-];
-
-const MockedOptions2 = ["0-18", "18-60"];
+import { ageOptions, eventOptions } from "../optionsData";
+import { useParams } from "react-router";
+import { useProjectForm } from "../../../../hooks/useProjectForm";
 
 export const AddProject = () => {
-  const formik = useFormik({
-    initialValues: {
-      title: "",
-      link: "",
-      description: "",
-      publishDate: null,
-      eventDate: null,
-      ageLimit: null,
-      category: null,
-      image: null,
-    },
-    validationSchema: Yup.object({
-      title: Yup.string().required("Обов'язкове поле"),
-      link: Yup.string().url("Невірний формат посилання"),
-      description: Yup.string().required("Обязательное поле"),
-      eventDate: Yup.string()
-        .matches(
-          /^(0[1-9]|1[0-2])\.\d{4} - (0[1-9]|1[0-2])\.\d{4}$/,
-          "Невірний формат",
-        )
-        .required("Обов'язкове поле"),
-    }),
-    onSubmit: (values) => {
-      if (!formik.values.publishDate) {
-        const currentDate = new Date();
-        formik.values.publishDate = `${String(
-          currentDate.getMonth() + 1,
-        ).padStart(2, "0")}.${currentDate.getFullYear()}`;
-      }
-
-      console.log("Sending data:", values);
-    },
-  });
+  const { id } = useParams();
+  const { formik } = useProjectForm(id);
 
   return (
     <>
@@ -110,12 +70,12 @@ export const AddProject = () => {
               error={formik.touched.eventDate && formik.errors.eventDate}
             />
             <CustomSelect
-              options={MockedOptions}
+              options={eventOptions}
               onChange={(option) => formik.setFieldValue("category", option)}
               error={formik.touched.category && formik.errors.category}
             />
             <CustomSelect
-              options={MockedOptions2}
+              options={ageOptions}
               onChange={(option) => formik.setFieldValue("ageLimit", option)}
               placeholder="Вікові обмеження"
               selectPrompt="Оберіть вік"
