+++ src/pages/AdminPanel/AdminProjects/optionsData.js
@@ -1,0 +1,8 @@
+export const eventOptions = [
+  "Тунельний політ",
+  "Банджі джампінг",
+  "Стрибки з парашутом",
+  "Стежка у хмарах",
+];
+
+export const ageOptions = ["0-18", "18-60"];
